requirements.txt
numpy
Pillow
opencv-python
fastapi
uvicorn
torch
tensorflow
ultralytics